+++ feature/collection/src/main/java/st/slex/csplashscreen/feature/collection/navigation/SingleCollectionRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.collection.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.collection.ui.store.SingleCollectionStore.Navigation
+import st.slex.csplashscreen.feature.collection.ui.presenter.SingleCollectionStore.Navigation
 
 interface SingleCollectionRouter : Router<Navigation>
