+++ feature/collection/src/main/java/st/slex/csplashscreen/feature/collection/navigation/SingleCollectionRouterImpl.kt
@@ -2,7 +2,7 @@ package st.slex.csplashscreen.feature.collection.navigation
 
 import st.slex.csplashscreen.core.navigation.NavigationScreen
 import st.slex.csplashscreen.core.ui.di.Navigator
-import st.slex.csplashscreen.feature.collection.ui.store.SingleCollectionStore.Navigation
+import st.slex.csplashscreen.feature.collection.ui.presenter.SingleCollectionStore.Navigation
 import javax.inject.Inject
 
 class SingleCollectionRouterImpl @Inject constructor(
