+++ feature/favourite/src/main/java/st/slex/csplashscreen/feature/favourite/di/FavouriteModule.kt
@@ -2,18 +2,16 @@ package st.slex.csplashscreen.feature.favourite.di
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
+import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import st.slex.csplashscreen.core.ui.di.ViewModelKey
 import st.slex.csplashscreen.feature.favourite.domain.FavouriteInteractor
 import st.slex.csplashscreen.feature.favourite.domain.FavouriteInteractorImpl
 import st.slex.csplashscreen.feature.favourite.navigation.FavouriteRouter
 import st.slex.csplashscreen.feature.favourite.navigation.FavouriteRouterImpl
-import st.slex.csplashscreen.feature.favourite.ui.FavouriteViewModel
-import st.slex.csplashscreen.feature.favourite.ui.store.FavouriteStore
-import st.slex.csplashscreen.feature.favourite.ui.store.FavouriteStoreImpl
+import st.slex.csplashscreen.feature.favourite.ui.presenter.FavouriteViewModel
 
 @Module
 interface FavouriteModule {
@@ -25,10 +23,6 @@ interface FavouriteModule {
     @Binds
     @FavouriteScope
     fun bindRouter(impl: FavouriteRouterImpl): FavouriteRouter
-
-    @Binds
-    @FavouriteScope
-    fun bindStore(impl: FavouriteStoreImpl): FavouriteStore
 
     @Binds
     @FavouriteScope
