+++ feature/favourite/src/main/java/st/slex/csplashscreen/feature/favourite/navigation/FavouriteRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.favourite.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.favourite.ui.store.FavouriteStore.Navigation
+import st.slex.csplashscreen.feature.favourite.ui.presenter.FavouriteStore.Navigation
 
 interface FavouriteRouter : Router<Navigation>
