+++ feature/home/src/main/java/st/slex/csplashscreen/feature/home/navigation/HomeRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.home.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.home.ui.store.HomeStore
+import st.slex.csplashscreen.feature.home.ui.presenter.HomeStore
 
 interface HomeRouter : Router<HomeStore.Navigation>
