+++ feature/home/src/main/java/st/slex/csplashscreen/feature/home/ui/component/tabs/MainScreenTabs.kt
@@ -1,4 +1,4 @@
-package st.slex.csplashscreen.feature.home.ui.components.tabs
+package st.slex.csplashscreen.feature.home.ui.component.tabs
 
 import androidx.annotation.StringRes
 import st.slex.csplashscreen.feature.home.R
