+++ feature/home/src/main/java/st/slex/csplashscreen/feature/home/ui/presenter/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package st.slex.csplashscreen.feature.home.ui.store
+package st.slex.csplashscreen.feature.home.ui.presenter
 
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
@@ -10,28 +10,28 @@ import st.slex.csplashscreen.core.collection.ui.model.toPresentation
 import st.slex.csplashscreen.core.core.coroutine.AppDispatcher
 import st.slex.csplashscreen.core.photos.ui.model.PhotoModel
 import st.slex.csplashscreen.core.photos.ui.model.toPresentation
-import st.slex.csplashscreen.core.ui.mvi.BaseStore
+import st.slex.csplashscreen.core.ui.mvi.BaseViewModel
 import st.slex.csplashscreen.core.ui.paging.PagingSource
 import st.slex.csplashscreen.feature.home.domain.HomeInteractor
 import st.slex.csplashscreen.feature.home.navigation.HomeRouter
-import st.slex.csplashscreen.feature.home.ui.store.HomeStore.Action
-import st.slex.csplashscreen.feature.home.ui.store.HomeStore.Event
-import st.slex.csplashscreen.feature.home.ui.store.HomeStore.Navigation
-import st.slex.csplashscreen.feature.home.ui.store.HomeStore.State
+import st.slex.csplashscreen.feature.home.ui.presenter.HomeStore.Action
+import st.slex.csplashscreen.feature.home.ui.presenter.HomeStore.Event
+import st.slex.csplashscreen.feature.home.ui.presenter.HomeStore.Navigation
+import st.slex.csplashscreen.feature.home.ui.presenter.HomeStore.State
 import javax.inject.Inject
 
-class HomeStoreImpl @Inject constructor(
+class HomeViewModel @Inject constructor(
     private val interactor: HomeInteractor,
     appDispatcher: AppDispatcher,
     router: HomeRouter
-) : HomeStore, BaseStore<State, Event, Action, Navigation>(router, appDispatcher) {
+) : BaseViewModel<State, Event, Action, Navigation>(router, appDispatcher) {
 
     override val initialState: State = State(
         collections = ::collections,
         photos = ::photos
     )
 
-    override val state: MutableStateFlow<State> = MutableStateFlow(initialState)
+    override val _state: MutableStateFlow<State> = MutableStateFlow(initialState)
 
     private val collections: StateFlow<PagingData<CollectionModel>>
         get() = Pager(config = config) { PagingSource(interactor::getAllCollections) }
@@ -41,7 +41,7 @@ class HomeStoreImpl @Inject constructor(
         get() = Pager(config = config) { PagingSource(interactor::getAllPhotos) }
             .state { image -> image.toPresentation() }
 
-    override fun processAction(action: Action) {
+    override fun sendAction(action: Action) {
         when (action) {
             is Action.OnCollectionClick -> actionCollectionClick(action)
             is Action.OnImageClick -> actionImageClick(action)
