+++ feature/photo-detail/src/main/java/st/slex/csplashscreen/feature/feature_photo_detail/di/ImageDetailModule.kt
@@ -2,18 +2,16 @@ package st.slex.csplashscreen.feature.feature_photo_detail.di
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
+import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import st.slex.csplashscreen.core.ui.di.ViewModelKey
 import st.slex.csplashscreen.feature.feature_photo_detail.domain.interactor.ImageDetailInteractor
 import st.slex.csplashscreen.feature.feature_photo_detail.domain.interactor.ImageDetailInteractorImpl
 import st.slex.csplashscreen.feature.feature_photo_detail.navigation.ImageDetailRouter
 import st.slex.csplashscreen.feature.feature_photo_detail.navigation.ImageDetailRouterImpl
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.ImageDetailViewModel
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.store.ImageDetailStore
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.store.ImageDetailStoreImpl
+import st.slex.csplashscreen.feature.feature_photo_detail.ui.presenter.ImageDetailViewModel
 import st.slex.csplashscreen.feature.feature_photo_detail.ui.utils.DownloadImageUseCase
 import st.slex.csplashscreen.feature.feature_photo_detail.ui.utils.DownloadImageUseCaseImpl
 import st.slex.csplashscreen.feature.feature_photo_detail.ui.utils.WallpaperSetUseCase
@@ -29,10 +27,6 @@ interface ImageDetailModule {
     @Binds
     @ImageDetailScope
     fun bindsRouter(impl: ImageDetailRouterImpl): ImageDetailRouter
-
-    @Binds
-    @ImageDetailScope
-    fun bindsStore(impl: ImageDetailStoreImpl): ImageDetailStore
 
     @Binds
     @ImageDetailScope
