+++ feature/photo-detail/src/main/java/st/slex/csplashscreen/feature/feature_photo_detail/navigation/ImageDetailRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.feature_photo_detail.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.store.ImageDetailStore.Navigation
+import st.slex.csplashscreen.feature.feature_photo_detail.ui.presenter.ImageDetailStore.Navigation
 
 interface ImageDetailRouter : Router<Navigation>
