+++ feature/photo-detail/src/main/java/st/slex/csplashscreen/feature/feature_photo_detail/navigation/ImageDetailRouterImpl.kt
@@ -2,7 +2,7 @@ package st.slex.csplashscreen.feature.feature_photo_detail.navigation
 
 import st.slex.csplashscreen.core.navigation.NavigationScreen
 import st.slex.csplashscreen.core.ui.di.Navigator
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.store.ImageDetailStore.Navigation
+import st.slex.csplashscreen.feature.feature_photo_detail.ui.presenter.ImageDetailStore.Navigation
 import javax.inject.Inject
 
 class ImageDetailRouterImpl @Inject constructor(
