+++ feature/photo-detail/src/main/java/st/slex/csplashscreen/feature/feature_photo_detail/ui/ImageDetailScreen.kt
@@ -39,12 +39,12 @@ import st.slex.csplashscreen.core.ui.components.UserImageHeadWithUserName
 import st.slex.csplashscreen.core.ui.theme.Dimen
 import st.slex.csplashscreen.feature.feature_photo_detail.R
 import st.slex.csplashscreen.feature.feature_photo_detail.domain.model.ImageDetail
+import st.slex.csplashscreen.feature.feature_photo_detail.ui.presenter.ImageDetailStore.State
 import st.slex.csplashscreen.feature.feature_photo_detail.ui.components.DetailImageBodyTags
-import st.slex.csplashscreen.feature.feature_photo_detail.ui.store.ImageDetailStore
 
 @Composable
 fun ImageDetailScreen(
-    state: ImageDetailStore.State,
+    state: State,
     onProfileClick: (String) -> Unit,
     onDownloadImageClick: () -> Unit,
     onTagClick: (String) -> Unit,
