+++ feature/photo-detail/src/main/java/st/slex/csplashscreen/feature/feature_photo_detail/ui/presenter/DownloadImageType.kt
@@ -1,4 +1,4 @@
-package st.slex.csplashscreen.feature.feature_photo_detail.ui.store
+package st.slex.csplashscreen.feature.feature_photo_detail.ui.presenter
 
 enum class DownloadImageType {
     ORIGINAL,
