+++ feature/search/src/main/java/st/slex/csplashscreen/feature/search/di/SearchPhotosModule.kt
@@ -2,10 +2,10 @@ package st.slex.csplashscreen.feature.search.di
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
+import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import st.slex.csplashscreen.core.ui.di.ViewModelKey
 import st.slex.csplashscreen.feature.search.data.SearchRepository
 import st.slex.csplashscreen.feature.search.data.SearchRepositoryImpl
@@ -13,9 +13,7 @@ import st.slex.csplashscreen.feature.search.domain.interactor.SearchPhotosIntera
 import st.slex.csplashscreen.feature.search.domain.interactor.SearchPhotosInteractorImpl
 import st.slex.csplashscreen.feature.search.navigation.SearchPhotosRouter
 import st.slex.csplashscreen.feature.search.navigation.SearchPhotosRouterImpl
-import st.slex.csplashscreen.feature.search.ui.SearchViewModel
-import st.slex.csplashscreen.feature.search.ui.store.SearchStore
-import st.slex.csplashscreen.feature.search.ui.store.SearchStoreImpl
+import st.slex.csplashscreen.feature.search.ui.presenter.SearchViewModel
 
 @Module
 interface SearchPhotosModule {
@@ -31,10 +29,6 @@ interface SearchPhotosModule {
     @Binds
     @SearchPhotosScope
     fun bindsRouter(impl: SearchPhotosRouterImpl): SearchPhotosRouter
-
-    @Binds
-    @SearchPhotosScope
-    fun bindsStore(impl: SearchStoreImpl): SearchStore
 
     @Binds
     @SearchPhotosScope
