+++ feature/search/src/main/java/st/slex/csplashscreen/feature/search/navigation/SearchPhotosRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.search.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.search.ui.store.SearchStore.Navigation
+import st.slex.csplashscreen.feature.search.ui.presenter.SearchStore.Navigation
 
 interface SearchPhotosRouter : Router<Navigation>
