+++ feature/search/src/main/java/st/slex/csplashscreen/feature/search/navigation/SearchPhotosRouterImpl.kt
@@ -2,7 +2,7 @@ package st.slex.csplashscreen.feature.search.navigation
 
 import st.slex.csplashscreen.core.navigation.NavigationScreen
 import st.slex.csplashscreen.core.ui.di.Navigator
-import st.slex.csplashscreen.feature.search.ui.store.SearchStore.Navigation
+import st.slex.csplashscreen.feature.search.ui.presenter.SearchStore.Navigation
 import javax.inject.Inject
 
 class SearchPhotosRouterImpl @Inject constructor(
