+++ feature/user/src/main/java/st/slex/csplashscreen/feature/user/di/UserModule.kt
@@ -2,10 +2,10 @@ package st.slex.csplashscreen.feature.user.di
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
+import st.slex.csplashscreen.core.ui.base.ViewModelFactory
 import st.slex.csplashscreen.core.ui.di.ViewModelKey
 import st.slex.csplashscreen.feature.user.data.UserRepository
 import st.slex.csplashscreen.feature.user.data.UserRepositoryImpl
@@ -13,9 +13,7 @@ import st.slex.csplashscreen.feature.user.domain.UserInteractor
 import st.slex.csplashscreen.feature.user.domain.UserInteractorImpl
 import st.slex.csplashscreen.feature.user.navigation.UserRouter
 import st.slex.csplashscreen.feature.user.navigation.UserRouterImpl
-import st.slex.csplashscreen.feature.user.ui.UserViewModel
-import st.slex.csplashscreen.feature.user.ui.store.UserStore
-import st.slex.csplashscreen.feature.user.ui.store.UserStoreImpl
+import st.slex.csplashscreen.feature.user.ui.presenter.UserViewModel
 
 @Module
 interface UserModule {
@@ -31,10 +29,6 @@ interface UserModule {
     @Binds
     @UserScope
     fun bindsRouter(impl: UserRouterImpl): UserRouter
-
-    @Binds
-    @UserScope
-    fun bindsStore(impl: UserStoreImpl): UserStore
 
     @Binds
     @IntoMap
