+++ feature/user/src/main/java/st/slex/csplashscreen/feature/user/navigation/UserRouter.kt
@@ -1,6 +1,6 @@
 package st.slex.csplashscreen.feature.user.navigation
 
 import st.slex.csplashscreen.core.ui.mvi.Router
-import st.slex.csplashscreen.feature.user.ui.store.UserStore
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore
 
 interface UserRouter : Router<UserStore.Navigation>
