+++ feature/user/src/main/java/st/slex/csplashscreen/feature/user/ui/UserScreen.kt
@@ -25,7 +25,7 @@ import st.slex.csplashscreen.feature.user.ui.components.pager.UserPager
 import st.slex.csplashscreen.feature.user.ui.components.toolbar.UserToolbar
 import st.slex.csplashscreen.feature.user.ui.state.UserPagerState
 import st.slex.csplashscreen.feature.user.ui.state.UserSwipeState
-import st.slex.csplashscreen.feature.user.ui.store.UserStore
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore
 import st.slex.csplashscreen.feature.user.ui.utils.SwipeState
 
 @OptIn(ExperimentalMaterialApi::class)
