--- feature/user/src/main/java/st/slex/csplashscreen/feature/user/ui/UserViewModel.kt
@@ -1,12 +1,0 @@
-package st.slex.csplashscreen.feature.user.ui
-
-import st.slex.csplashscreen.core.ui.base.BaseViewModel
-import st.slex.csplashscreen.feature.user.ui.store.UserStore
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.Action
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.Event
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.State
-import javax.inject.Inject
-
-class UserViewModel @Inject constructor(
-    store: UserStore
-) : BaseViewModel<State, Event, Action>(store)
