+++ feature/user/src/main/java/st/slex/csplashscreen/feature/user/ui/presenter/UserViewModel.kt
@@ -1,4 +1,4 @@
-package st.slex.csplashscreen.feature.user.ui.store
+package st.slex.csplashscreen.feature.user.ui.presenter
 
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
@@ -10,21 +10,21 @@ import st.slex.csplashscreen.core.collection.ui.model.toPresentation
 import st.slex.csplashscreen.core.core.coroutine.AppDispatcher
 import st.slex.csplashscreen.core.photos.ui.model.PhotoModel
 import st.slex.csplashscreen.core.photos.ui.model.toPresentation
-import st.slex.csplashscreen.core.ui.mvi.BaseStore
+import st.slex.csplashscreen.core.ui.mvi.BaseViewModel
 import st.slex.csplashscreen.core.ui.paging.PagingSource
 import st.slex.csplashscreen.feature.user.domain.UserInteractor
 import st.slex.csplashscreen.feature.user.navigation.UserRouter
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.Action
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.Event
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.Navigation
-import st.slex.csplashscreen.feature.user.ui.store.UserStore.State
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore.Action
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore.Event
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore.Navigation
+import st.slex.csplashscreen.feature.user.ui.presenter.UserStore.State
 import javax.inject.Inject
 
-class UserStoreImpl @Inject constructor(
+class UserViewModel @Inject constructor(
     private val interactor: UserInteractor,
     appDispatcher: AppDispatcher,
     router: UserRouter
-) : UserStore, BaseStore<State, Event, Action, Navigation>(router, appDispatcher) {
+) : BaseViewModel<State, Event, Action, Navigation>(router, appDispatcher) {
 
     override val initialState: State = State(
         user = null,
@@ -33,9 +33,9 @@ class UserStoreImpl @Inject constructor(
         collections = ::getCollections
     )
 
-    override val state: MutableStateFlow<State> = MutableStateFlow(initialState)
+    override val _state: MutableStateFlow<State> = MutableStateFlow(initialState)
 
-    override fun processAction(action: Action) {
+    override fun sendAction(action: Action) {
         when (action) {
             is Action.Init -> actionInit(action)
             is Action.OnBackButtonClick -> actionBackClick()
